init GrammarChecker struct
