+++ JSONParser/JSONParser/Analysis.swift
@@ -87,68 +87,5 @@ struct Analysis {
             return dictionary
         }
     }
-    
-    // [] or {} 쌍 검사
-    public static func checkPair(to jsonData:String) -> Bool {
-        let countPrefix1 = jsonData.contains("{")
-        let countPrefix2 = jsonData.contains("[")
-        let countSuffix1 = jsonData.contains("}")
-        let countSuffix2 = jsonData.contains("]")
-        guard countPrefix1 == countSuffix1 && countPrefix2 == countSuffix2 else { return false }
-        return true
-    }
-    
-    // [] or {} 패턴 검사
-    public static func checkPattern(to jsonData:String) -> Bool {
-        var data = jsonData
-        
-        data.removeFirst()
-        data.removeLast()
-        
-        var before = ""
-        /*
-         [{{,}}]
-         {[[,]]}
-         {][[[]]}
-         [}{]
-         {[],[]}
-         */
-        for d in data {
-            // 3 : } or ] 이게 처음부터 나오면 형태 잘못된 것 입니다.
-            if before.isEmpty && (d == "}" || d == "]"){
-                return false
-            }
-            // 1 : { or [ 이면 before에 값을 저장함
-            if d == "{" || d == "[" {
-                before = String(d)
-            }
-            // 2 : ] or } 이면 before에 동일한 형태가 있는지 확인
-            if d == "}" || d == "]" {
-                if d == "}" && before == "{" {
-                    // Pass : before 초기화
-                    before = ""
-                }else if d == "]" && before == "[" {
-                    // Pass : before 초기화
-                    before = ""
-                }else {
-                    return false
-                }
-            }
-        }
-        return true
-    }
+
 }
-
-extension String {
-    func contains(_ find: String) -> Int{
-        var count = 0
-        for char in self {
-            if find.contains(char) {
-                count = count + 1
-            }
-        }
-        return count
-    }
-}
-
-
